requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
opencv-python
ultralytics
paddleocr
numpy
requests
cvzone
